requirements.txt
duckdb==1.1.3
pandas==2.2.2
numpy==2.0.1
pyarrow==17.0.0
pydantic==2.8.2
python-dotenv==1.0.1
rich==13.7.1
streamlit==1.37.0
matplotlib==3.9.0

# dbt (local)
dbt-core==1.8.6
dbt-duckdb==1.8.2

# dev
ruff==0.6.2
pytest==8.3.2
